# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Tonatiuh Trading SaaS API
  version: 1.0.0
  description: Multi-tenant API for accounts, subscriptions, exchange connections, and trading bots.
servers:
  - url: https://api.example.com
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/exchanges:
    get:
      tags: [Exchanges]
      operationId: listExchangeConnections
      responses:
        '200': {description: Connections without credentials}
        default: {$ref: '#/components/responses/Error'}
    post:
      tags: [Exchanges]
      operationId: createExchangeConnection
      requestBody: {$ref: '#/components/requestBodies/ExchangeConnection'}
      responses:
        '201': {description: Encrypted connection created}
        default: {$ref: '#/components/responses/Error'}
  /api/v1/exchanges/{id}:
    patch:
      tags: [Exchanges]
      operationId: updateExchangeConnection
      parameters: [{$ref: '#/components/parameters/ResourceId'}]
      requestBody: {$ref: '#/components/requestBodies/ExchangeConnectionPatch'}
      responses:
        '200': {description: Connection updated}
        default: {$ref: '#/components/responses/Error'}
  /api/v1/exchanges/{id}/verify:
    post:
      tags: [Exchanges]
      operationId: verifyExchangeConnection
      parameters: [{$ref: '#/components/parameters/ResourceId'}]
      responses:
        '200': {description: Credentials verified through the exchange API}
        default: {$ref: '#/components/responses/Error'}
  /api/v1/bots:
    get:
      tags: [Bots]
      operationId: listBots
      responses:
        '200': {description: Organization bots}
        default: {$ref: '#/components/responses/Error'}
    post:
      tags: [Bots]
      operationId: createBot
      requestBody: {$ref: '#/components/requestBodies/Bot'}
      responses:
        '201': {description: Bot created}
        default: {$ref: '#/components/responses/Error'}
  /api/v1/bots/{id}:
    get:
      tags: [Bots]
      operationId: getBot
      parameters: [{$ref: '#/components/parameters/ResourceId'}]
      responses:
        '200': {description: Bot details}
        default: {$ref: '#/components/responses/Error'}
    patch:
      tags: [Bots]
      operationId: updateBot
      parameters: [{$ref: '#/components/parameters/ResourceId'}]
      requestBody: {$ref: '#/components/requestBodies/BotPatch'}
      responses:
        '200': {description: Stopped bot updated}
        default: {$ref: '#/components/responses/Error'}
...
  /api/v1/bots/{id}/commands:
    get:
      tags: [Bots]
      operationId: listBotCommands
      parameters: [{$ref: '#/components/parameters/ResourceId'}]
      responses:
        '200': {description: Last 100 commands}
        default: {$ref: '#/components/responses/Error'}
  /api/v1/bots/{id}/orders:
    get:
      tags: [Bots]
      operationId: listBotOrders
      parameters: [{$ref: '#/components/parameters/ResourceId'}]
      responses:
        '200': {description: Last 200 orders}
        default: {$ref: '#/components/responses/Error'}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
    metricsToken: {type: http, scheme: bearer, bearerFormat: opaque metrics token}
    emailWebhookToken: {type: http, scheme: bearer, bearerFormat: opaque provider webhook token}
  parameters:
    ResourceId: {name: id, in: path, required: true, schema: {type: string, format: uuid}}
...
  responses:
    Error:
      description: Standard API error
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}
...
  requestBodies:
...
    ExchangeConnection:
      required: true
      content: {application/json: {schema: {$ref: '#/components/schemas/ExchangeConnectionInput'}}}
    ExchangeConnectionPatch:
      required: true
      content: {application/json: {schema: {type: object, additionalProperties: false, properties: {label: {type: string}, enabled: {type: boolean}, apiKey: {type: string, writeOnly: true}, secret: {type: string, writeOnly: true}, password: {type: string, writeOnly: true}}}}}
    Bot:
      required: true
      content: {application/json: {schema: {$ref: '#/components/schemas/BotInput'}}}
    BotPatch:
      required: true
      content: {application/json: {schema: {type: object, additionalProperties: false, properties: {name: {type: string}, configuration: {$ref: '#/components/schemas/TradingConfiguration'}}}}}
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties: {code: {type: string}, message: {type: string}, requestId: {type: string}, details: {}}
...
    ExchangeConnectionInput:
      type: object
      additionalProperties: false
      required: [exchange, label, apiKey, secret]
      properties: {exchange: {type: string, enum: [okx, binance, bitget, kucoin, mexc, poloniex, gate, exmo, bybit]}, label: {type: string}, apiKey: {type: string, writeOnly: true}, secret: {type: string, writeOnly: true}, password: {type: string, writeOnly: true}, sandbox: {type: boolean}, verify: {type: boolean}}
    BotInput:
      type: object
      additionalProperties: false
      required: [exchangeConnectionId, name, configuration]
      properties: {exchangeConnectionId: {type: string, format: uuid}, name: {type: string}, strategy: {type: string, enum: [VECTOR_PROFIT]}, configuration: {$ref: '#/components/schemas/TradingConfiguration'}}
    TradingConfiguration:
      type: object
      additionalProperties: false
      required: [symbol]
      properties:
        symbol: {type: string, pattern: '^[A-Z0-9]{2,12}/[A-Z0-9]{2,12}$'}
        positionSize: {type: number, exclusiveMinimum: 0}
        countGridSize: {type: number, minimum: 1, maximum: 100}
        gridSize: {type: number, exclusiveMinimum: 0}
        percentBuyBackStep: {type: number, minimum: 0, maximum: 1}
        takeProfit: {type: number, minimum: 0, maximum: 1}
        stopLoss: {type: number, minimum: 0, maximum: 1}
        percentProfit: {type: number, minimum: 0, maximum: 1}
        percentFromBalance: {type: number, exclusiveMinimum: 0, maximum: 1}
        candlePriceRange: {type: string, enum: [1m, 5m, 15m, 1h, 4h, 1d]}
        isFibonacci: {type: boolean}
        isPercentTargetAfterTakeProfit: {type: boolean}
        isCapitalizeDeltaFromSale: {type: boolean}
        isCoinAccumulation: {type: boolean}
        isOnlyBuy: {type: boolean}
        percentTargetAfterTakeProfit: {type: number, minimum: 0, maximum: 1}
        balanceDistribution: {type: boolean}